make useTranslation wrapper android-only
